Add plumbing for indirect output plugins
An indirect output plugin does not have a flush callback, but has a
get_output callback that returns an output instance that should be used
for flushing. We call this callback when creating flush tasks to get
them routed to the correct output.

This allows for the creation of output plugins that decide dynamically
which output to use. An example of this would be an output plugin that
shards logs over multiple kafka streams.